requirements.txt
python-telegram-bot==20.1
httpx==0.24.0
python-dotenv==1.0.0
